refactor: 상태 및 핸들러 훅 통합
